requirements.txt
pyyaml==6.0.2
requests==2.32.3
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
pytest==8.3.3
prometheus-client==0.20.0
httpx==0.27.2
starlette==0.40.0